A simple template to create IAM user
This template creates an IAM user with three IAM policies attached to it. The first policy is an AWS managed policy called ReadOnlyAccess. The second policy is a customer managed policy called my_customer_managed_policy and this policy is already present in the account(s) where the stack(s) will be created. Since the policy ARN for the customer managed policy in different accounts differ by the account ID, the pseudo parameter AWS::AccountId is used. The full policy ARN can then be obtained by using the function Join. The third policy is an inline policy called demo-policy. The outputs of the stack will show the values of the access key and secret access key for the IAM user.