Classic load balancer with logging enabled
This template can be used to create a classic elastic load balancer (ELB) in three different regions - North Virginia, Ireland, and Singapore. If the ELB is created in the North Virginia, the ELB logs will be sent to an existing S3 bucket in the AWS account that is also located in the North Virginia region. Likewise for ELB in the other two regions. The template also creates a security group that will be attached to the ELB.